Initial Commit
Work completed had shared communication device that was accessible
through AppDelegate. This allowed the calls to be made (based for
long-runners) and for screens to be aware if calls were either
finished, or still executing. The screens that are fully functional in
this commit are:
- Directory Search
- Accounts
- Schedule
- Recreation (which used an API created that was intended for campus
recreation to utilize that was never followed through on)
- Settings (although the page was laid out, all calls use hard coded
values for testing)